% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{amphibiaBRlist}
\alias{amphibiaBRlist}
\title{Brazilian Red List of Amphibians species}
\format{
\subsection{\code{amphibiaBRlist}}{

A data frame with 61 rows and 5 columns:
\describe{
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{species}{Scientific name}
\item{category}{Category of threat}
\item{MMA2014}{Species was evaluated in the past list (2014)}
...
}
}
}
\usage{
amphibiaBRlist
}
\description{
Dataset of the official Brazilian Threatened Species List
published by the Ministério do Meio Ambiente
}
\details{
Source availabe at \href{https://www.in.gov.br/en/web/dou/-/portaria-gm/mma-n-300-de-13-de-dezembro-de-2022-450425464}{MMA 300/2022}
}
\keyword{datasets}
