% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{plantsBRlist}
\alias{plantsBRlist}
\title{Brazilian Red List of Plant species}
\format{
\subsection{\code{plantsBRlist}}{

A data frame with 61 rows and 5 columns:
\describe{
\item{order}{Taxonomic Order}
\item{family}{Taxonomic Family}
\item{species}{Scientific name}
\item{category}{Category of threat}
\item{MMA2014}{Species was evaluated in the past list (2014)}
...
}
}
}
\usage{
plantsBRlist
}
\description{
Dataset of the official Brazilian Threatened Species List
published by the Ministério do Meio Ambiente
}
\details{
Source availabe at \href{https://www.in.gov.br/web/dou/-/portaria-mma-n-148-de-7-de-junho-de-2022-406272733}{MMA 148/2022}
}
\keyword{datasets}
